+++ src/test/java/com/project/kodesalon/model/member/service/MemberServiceTest.java
@@ -1,8 +1,8 @@
 package com.project.kodesalon.model.member.service;
 
 import com.project.kodesalon.model.member.domain.Member;
-import com.project.kodesalon.model.member.dto.LoginRequestDto;
-import com.project.kodesalon.model.member.dto.LoginResponseDto;
+import com.project.kodesalon.model.member.service.dto.LoginRequestDto;
+import com.project.kodesalon.model.member.service.dto.LoginResponseDto;
 import com.project.kodesalon.model.member.repository.MemberRepository;
 import org.assertj.core.api.BDDSoftAssertions;
 import org.junit.jupiter.api.DisplayName;
